Stop calling `rb_obj_id` on T_IMEMO
In Ruby 3.5 object_id is now an inline property of objects and
T_IMEMO can't have fields.

But this code was written in a time when `rb_obj_id` wasn't much
more than boxing the object address in a Fixnum, hence we can
just explictly do that with `ULL2NUM`.